Refactor post & tax contracts and traits
Fixes #21
