+++ lib/Model/PostTypes/Example.php
@@ -7,7 +7,9 @@
 
 namespace Pixels\ProjectName\Model\PostTypes;
 
-use Pixels\ProjectName\Model\TraitSingleton;
+// Contracts.
+use Pixels\ProjectName\Model\PostTypes\Contracts\AbstractPostType;
+use Pixels\ProjectName\Model\PostTypes\Contracts\PostTypeInterface;
 
 /**
  * Register Example class
@@ -21,9 +23,6 @@ class Example extends AbstractPostType implements PostTypeInterface {
 	 * --> If false, autocreate labels from one word
 	 */
 	const TRANSLATE_LABELS = false;
-
-	// Trait that allows singleton behavior.
-	use TraitSingleton;
 
 	/**
 	 * Class constructor
@@ -121,4 +120,4 @@ class Example extends AbstractPostType implements PostTypeInterface {
 
 		return $labels;
 	}
-} //end Example
+}
